+++ src/main/webapp/components/query-status/index.js
@@ -1,0 +1,2 @@
+import QueryStatus from './query-status'
+export { QueryStatus as default }
