+++ src/main/webapp/components/query-status/query-status.js
